requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
argcomplete==1.12.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bokeh==2.3.3
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
discord.py==1.7.3
entrypoints==0.3
fer==21.0.3
firebase-admin==5.0.2
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-api-core==1.31.2
google-api-python-client==2.17.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.0.0
google-cloud-firestore==2.2.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.1.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.4
ipykernel==6.3.0
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==7.0.2
jupyter-core==4.7.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
livelossplot==0.5.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
msgpack==1.0.2
mtcnn==0.1.1
multidict==5.1.0
nest-asyncio==1.5.1
nltk==3.6.2
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.2.1
qrcode==7.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
textblob==0.15.3
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
utils==1.0.1
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
